Abstract cart cookie handling to a common lib
